project-ampule.code-workspace/requirements.txt
faiss-cpu
sentence-transformers
llama-cpp-python
numpy
requests
PyPDF2
beautifulsoup4
